pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "murban-copilot"
version = "0.1.0"
description = "AI-powered Murban-Brent crude oil spread analysis"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "plotly>=5.0",
    "streamlit>=1.28",
    "yfinance>=0.2.30",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = [
    "llama-cpp-python>=0.2.0",
    "huggingface_hub>=0.19",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "freezegun>=1.2",
]

[tool.setuptools.packages.find]
where = ["src"]
